% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Animal}
\alias{Animal}
\title{Animal with capacity for attraction to other latent animals}
\arguments{
\item{init_latent}{latent animals to be attracted to}

\item{init_animal}{index of latent animal to use as starting location}
}
\description{
Initialize an animal with latent members, and couple the observed position to one of the latent animals
}
