% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cop_threshold}
\alias{cop_threshold}
\title{Feeding effort as function of copepod density}
\usage{
cop_threshold(gamma, D)
}
\arguments{
\item{gamma}{Feeding threshold}

\item{D}{Coepepod density}
}
\description{
Feeding effort as function of copepod density
}
