% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deg2radians}
\alias{deg2radians}
\title{Convert degrees to radians}
\usage{
deg2radians(angle)
}
\arguments{
\item{angle}{Input angle in degrees}
}
\value{
Equivalent angle in radians
}
\description{
Convert degrees to radians
}
