% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narwind-package.R
\docType{data}
\name{density_narw}
\alias{density_narw}
\title{North Atlantic right whale density}
\format{
A list of twelve \code{SpatialGridDataFrame}
\describe{
  \item{Jan}{Density surface for January}
  \item{Feb}{Density surface for February}
  \item{Mar}{Density surface for March}
  \item{Apr}{Density surface for April}
  \item{May}{Density surface for May}
  \item{Jun}{Density surface for June}
  \item{Jul}{Density surface for July}
  \item{Aug}{Density surface for August}
  \item{Sep}{Density surface for September}
  \item{Oct}{Density surface for October}
  \item{Nov}{Density surface for November}
  \item{Dec}{Density surface for December}
}
}
\source{
Data provided by Jason Roberts (Marine Geospatial Ecology Lab, Duke University).
}
\description{
Data from .
}
\references{
Smith J (2006) Title, 41 p.
}
\keyword{datasets}
