% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.narwsim.R
\name{export.narwsim}
\alias{export.narwsim}
\title{Export simulation data}
\usage{
\method{export}{narwsim}(obj, prefix = "narwsim", ...)
}
\arguments{
\item{obj}{An object of class \code{narwsim}, as returned by \code{\link{narw}}.}

\item{prefix}{Character string. Output file name. Defaults to \code{"narwsim"}.}

\item{...}{Additional arguments passed to \code{\link[openxlsx]{write.xlsx}}.}
}
\value{
One Excel file per population cohort. It is recommended to only use this function for short simulation runs (\code{nsim = 100} or less, or to use the \code{cohort} and \code{whale} arguments to extract data for specific individuals/cohorts. Each output file contains multiple sheets capturing all simulation parameters, as follows:
\itemize{
\item \code{Attributes}: day, date, month, unique whale ID, cohort ID, alive (yes/no), age, starve (yes/no), mortality from other sources (yes/no), probability of death from other sources, body condition, body length, coefficients of the length-at-age function, total mass, lean mass, fat mass, coefficients of the mass-at-length function, mouth radius, mouth opening angle, mouth width, migration to SEUS (yes/no), migration to GSL (yes/no)
\item \code{Attributes_calf} same as \code{Attributes}, for calves
\item \code{Movements}: easting,	northing, region, distance traveled, swim speed
\item \code{Behavior}: proportion of time spent gliding, proportion of time spent gliding during foraging, proportion of time spent gliding while swimming in echelon position, time spent traveling, time spent feeding, time spent resting/nursing
\item \code{Stressors}: Probability of entanglement, daily energetic cost of entanglement, entanglement event (yes/no), site of attachement (head/not head), entanglement serverity, entanglement duration, day of entanglement start, day of entanglement end, vessel strike	risk,	incidence of strike, incidence of behavioral response to noise, received sound level, threshold of response
\item \code{Energy}: Daily energy balance, daily energy intake, daily energy expenditure, resting metabolic rate, locomotory costs, scalar applied to locomotory costs, stroke rate, stroke rate during foraging, energetic cost of somatic growth, total energetic cost of gestation, energetic costs of fetus growth and placental maintenance, heat increment of gestation, energetic cost of lactation, relative proportions of bones, viscera and muscles in lean mass
\item \code{Energy_calf} same as \code{Energy}, for calves (where relevant)
\item \code{Feeding}: incidence of foraging, prey concentration, minimum prey threshold that triggers foraging, size of the mouth gape, swimming speed while foraging, prey capture efficiency, reduction in the mouth gape during a head entanglement, feeding effort, feed_effort, coefficients of the feeding effort curve, target body condition, average mass of copepods, energy content of copepods, digestive efficiency, metabolizing efficiency	for juveniles and adults (heat increment of feeding)
\item \code{Growth}: change in fat mass, daily change in lean mass, anabolism efficiency, catabolism efficiency
\item \code{Growth_calf} same as \code{Growth}, for calves
\item \code{Gestation}: incidence of abortion, fetus length, fetus mass, change in fetus mass, change in fetus length, predicted birth length, predicted birth mass, mass of muscles, viscera, bones, and blubber in fetal tissues, proportions of lipid and protein in fetal muscles, viscera, bones, bones, and blubber.
\item \code{Lactation}: duration of lactation, milk assimilation ratem milk provision rate, factor defining the nonlinearity between milk supply and body condition of the mother, age at which milk consumption starts to decrease, factor defining the nonlinearity between milk assimilation and calf age, mass of mammary glands, mammary efficiency, milk production rate, time spent nursing, target body condition (calf), proportions of protein and lipids in milk

}
}
\description{
Writes outputs from the individual-based model to an .xlsx file on disk.
}
\note{
This function
}
\examples{
\dontrun{
library(narwind)
m <- narw(1000)
export(m)}
}
\author{
Phil J. Bouchet
}
