% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feeding_effort}
\alias{feeding_effort}
\title{Title}
\usage{
feeding_effort(eta, target_condition = 0.3, maintenance_mass, blubber_mass)
}
\arguments{
\item{eta}{Steepness of the non-linear relationship}

\item{target_condition}{}

\item{maintenance_mass}{}

\item{blubber_mass}{}
}
\description{
Title
}
