% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{feeding_threshold_vec}
\alias{feeding_threshold_vec}
\title{Incidence of foraging behavior (vectorized)}
\usage{
feeding_threshold_vec(min_prey, prey_concentration)
}
\arguments{
\item{min_prey}{Minimum prey density threshold that triggers foraging}

\item{D}{Prey concentration}
}
\description{
Determines whether the prey concentration encountered by an animal
is sufficient to support foraging
}
