% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fetal_mass (vectorised)}
\alias{fetal_mass (vectorised)}
\alias{fetal_mass_vec}
\title{Fetal mass}
\usage{
fetal_mass_vec(days_to_birth, mother_length, bbc = 0L, body_density = 805.07)
}
\arguments{
\item{days_to_birth}{Number of days until birth, assuming a 365-day gestation period (d)}

\item{mother_length}{Body length of the mother (m)}

\item{bbc}{Body condition, as defined by Christiansen et al. (2022). Defaults to 0 for an individual of average condition.}

\item{body_density}{Average body density (kg/m^3)}
}
\description{
Fetal mass
}
\note{
In this parameterization, birth corresponds to t=0 and conception corresponds to t=-365
}
