% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mortality}
\alias{find_mortality}
\title{Estimate daily p(dead) to match a target annual mortality rate}
\usage{
find_mortality(annual.rate = 0.02)
}
\arguments{
\item{annual.rate}{Double. Target annual mortality rate (in %)}
}
\value{
Daily mortality probability.
}
\description{
Estimate daily p(dead) to match a target annual mortality rate
}
\note{
If the daily probability of a mortality event occurring is p, then the daily probability of
survival is (1-p). The probability of survival over a year is then given by (1-p)^365.
Therefore the probability of a mortality event occurring at least once during the year
(annual mortality rate) is p_year = 1-((1-p)^365). It follows that p_day = 1-((1-p_year)^(1/365))
}
