% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt_chart.R
\name{gantt_chart}
\alias{gantt_chart}
\title{Schedule of piling activities}
\usage{
gantt_chart(obj)
}
\arguments{
\item{obj}{An object of class \code{narwscenario}.}
}
\value{
A ggplot2 object
}
\description{
Plots a gantt chart showing the
}
\author{
Phil J. Bouchet
}
