% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ais}
\alias{get_ais}
\title{Download density maps}
\usage{
get_ais(year = 2021)
}
\arguments{
\item{version}{Version of the North Atlantic right whale model}
}
\description{
Download density maps from the Duke University website and unzip files
}
