% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_density}
\alias{get_density}
\title{Download density maps}
\usage{
get_density(region = NULL, version = 12, start = 2010, end = 2019)
}
\arguments{
\item{version}{Version of the North Atlantic right whale model}
}
\description{
Download density maps from the Duke University website and unzip files
}
