% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prey}
\alias{get_prey}
\title{Prey surfaces}
\usage{
get_prey(months = 1:12, do.plot = TRUE)
}
\description{
Generates monthly prey abundance rasters from the DFO data
}
\details{
------------------------------------------
}
