% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{growth_cost}
\alias{growth_cost}
\alias{growth_cost_old}
\title{Energetic cost of growth}
\usage{
growth_cost_old(
  delta_m,
  prop_blubber,
  prop_water,
  P_lipid_blubber,
  rho_lipid,
  rho_protein,
  D_lipid,
  D_protein
)
}
\arguments{
\item{delta_m}{Body mass growth increment (kg/day)}

\item{prop_blubber}{Proportion of the body that is blubber (\%)}

\item{prop_water}{Proportion of lean body mass that is water (\%)}

\item{P_lipid_blubber}{Proportion of blubber that is lipid (\%)}

\item{rho_lipid}{Energy density of lipids (kJ/kg)}

\item{rho_protein}{Energy density of protein (kJ/kg)}

\item{D_lipid}{Efficiency of deposition of lipids (\%)}

\item{D_protein}{Efficiency of deposition of protein (\%)}
}
\description{
Energetic cost of growth
}
