% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{locomotor_costs}
\alias{locomotor_costs}
\title{Costs of locomotion}
\usage{
locomotor_costs(
  mass,
  distance,
  strokerate_foraging,
  strokerate,
  glide_foraging,
  glide,
  t_feed,
  t_activ,
  scalar
)
}
\description{
Costs of locomotion
}
