% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_noise.R
\name{map_noise}
\alias{map_noise}
\title{Noise levels}
\usage{
map_noise(obj = NULL)
}
\arguments{
\item{obj}{An object of class \code{narwscenario}.}
}
\value{
A list containing noise layers for each day, in \code{SpatialGridDataFrame} format.
}
\description{
Generates rasters describing the daily noise footprint of selected wind farms, as defined in a scenario object.
}
\note{
The function relies on a simple propagation model which assumes that transmission loss depends on both log-range and frequency-specific absorption. The model is of the form: TL = b x log10R + a x R, where b = 15 and a = 1.175 by default. The resultant sound pressure level (SPL) from multiple sources is calculated based on the logarithmic addition of sound intensities from each source.
}
\examples{
\dontrun{
noisemaps <- map_noise(obj = scenario_01)
}
}
\author{
Phil J. Bouchet
}
