% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_vessels.R
\name{map_vessels}
\alias{map_vessels}
\title{Vessel strikes}
\usage{
map_vessels(
  obj = NULL,
  z = "dist",
  strike_scalar = 1e-07,
  which.month = 1:12,
  vessel.speed = NULL,
  speed.limit = NULL,
  baseline = TRUE,
  do.plot = FALSE,
  spgdf = FALSE
)
}
\arguments{
\item{obj}{An object of class \code{narwscenario}.}

\item{z}{Character. Output variable; can be one of "risk" (relative strike risk), "density" (transit density), "transits" (total number of transits per cell), or "dist" (total distance traveled per cell).}

\item{strike_scalar}{Numeric. Value of the factor used to scale the resulting strike surface(s).}

\item{which.month}{Integer, or vector of integers. Months of the year for which strike rasters are returned; 1 corresponds to January, 2 to February etc.}

\item{vessel.speed}{Numeric vector. Nominal transit speeds for each vessel class (Cable Lay, Construction/Crane, Crew Transfer,Heavy Cargo, Support Vessels <100 m, Survey, Tugs).}

\item{speed.limit}{Numeric vector of length 12, indicating the maximum speeds imposed during slow-downs for each month of the year. Months with \code{NA} values are subject to no speed restrictions.}

\item{baseline}{Logical. Defaults to \code{FALSE}. If \code{TRUE}, baseline conditions (excluding wind farm development activities) are included in the output.}

\item{do.plot}{Logical. Defaults to \code{FALSE}. If \code{TRUE}, strike maps are generated.}

\item{spgdf}{Defaults to \code{FALSE}. If \code{TRUE}, output surfaces are returned as \code{SpatialGridDataFrame} objects.}
}
\value{
A list containing strike risk rasters for each month, in \code{SpatialGridDataFrame} or \code{raster} format.
}
\description{
Generates monthly rasters describing the risk of vessel strikes from wind farm development activities, as defined in a scenario object.
}
\examples{
\dontrun{
library(narwind)
vesselmaps <- map_vessels(obj = scenario_01,
                          z = "dist",
                          strike_scalar = 1e-7,
                          which.month = 1:12,
                          vessel.speed = NULL,
                          speed.limit = c(10,10,10,10,rep(NA,8)),
                          baseline = TRUE,
                          gantt = TRUE,
                          do.plot = TRUE,
                          spgdf = FALSE)
}
...............................
...............................
...............................
...............................
...............................
...............................
}
\author{
Phil J. Bouchet
}
