% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mask_density}
\alias{mask_density}
\title{Clip and fill input raster}
\usage{
mask_density(x, mask, fill, offset = 0, clamp = FALSE)
}
\arguments{
\item{x}{Source raster to fill/pad.}

\item{mask}{Extent of the spatial region to fill.}

\item{fill}{Value used for filling.}

\item{offset}{Additional value to add to all locations within mask.}

\item{clamp}{Logical. Set to TRUE to give all locations outside mask region an NA value.}
}
\description{
Fill NA values within a mask region
}
