% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{milk_assimilation}
\alias{milk_assimilation}
\title{Decrease in milk feeding efficiency as a function of calf age}
\usage{
milk_assimilation(a = seq(0, 500), Tstop = 365, Tdecrease = 100, E = 0.9)
}
\arguments{
\item{a}{Calf age in days.}

\item{Tstop}{Time at which milk stops nursing and becomes nutritionally independent (in days).}

\item{Tdecrease}{Time at which milk comsuption starts to decrease (in days).}

\item{E}{Steepness of the decline}
}
\description{
Decrease in milk feeding efficiency as a function of calf age
}
