% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{milk_provisioning}
\alias{milk_provisioning}
\title{Increase in milk provisioning by female as a function of her body condition.}
\usage{
milk_provisioning(
  E = -2,
  blubber_mass,
  maintenance_mass,
  target_condition = 0.3,
  starvation = 0.15
)
}
\arguments{
\item{E}{Steepness of the non-linear relationship.}

\item{blubber_mass}{Mass of blubber (kg).}

\item{maintenance_mass}{Mass of structural tissues (kg).}

\item{target_condition}{Target body condition, expressed as the ratio of reserve to maintenance mass.}

\item{starvation}{Starvation threshold}
}
\description{
Increase in milk provisioning by female as a function of her body condition.
}
