% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NARW_simulator}
\alias{NARW_simulator}
\title{Simulate right whale movements}
\usage{
NARW_simulator(
  cohortID,
  seus,
  gsl,
  support,
  densities,
  densities_seus,
  densities_gsl,
  densitySeq,
  latentDensitySeq,
  prey,
  fishing,
  vessels,
  noise,
  doseresp_seed,
  doseresp,
  regions,
  limits,
  limits_regions,
  limits_prey,
  limits_fishing,
  limits_vessels,
  limits_noise,
  resolution,
  resolution_regions,
  resolution_prey,
  resolution_fishing,
  resolution_vessels,
  resolution_noise,
  M,
  stepsize,
  xinit,
  yinit,
  stressors,
  growth,
  prey_scale,
  starvation_df,
  starvation_death,
  starvation_onset,
  nursing_cessation,
  piling_hrs,
  progress
)
}
\arguments{
\item{densities}{array of density maps, stored in column-major order with 
indices [xind, yind, map] for projected spaces (xind is index for a projected lon coordinate, 
yind is index for a projected lat)}

\item{densitySeq}{1-indexed vector specifying which density maps should be
used to simulate movement at each timepoint; number of timepoints to
simulate comes from length of vector densitySeq}

\item{latentDensitySeq}{1-indexed vector specifying which density maps should be used to simulate movement for each latent animal}

\item{limits}{vector (xmin, xmax, ymin, ymax) of spatial coordinate extents for spatial densities}

\item{resolution}{vector (xres, yres) of spatial step sizes for spatial densities}

\item{M}{Number of proposals used in the importance sampler for movement (Michelot, 2019)}

\item{stepsize}{Rradius of the proposal circle for movement (Michelot, 2019)}

\item{xinit}{matrix of initial x coordinates for latent animals (nlatent x n)}

\item{yinit}{matrix of initial y coordinates for latent animals (nlatent x n)}
}
\value{
List of coordinates
}
\description{
Simulate right whale movements
}
