% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.narwproj.R
\name{plot.narwproj}
\alias{plot.narwproj}
\title{Plot population trends}
\usage{
\method{plot}{narwproj}(
  ...,
  diff = FALSE,
  interval = TRUE,
  shade = TRUE,
  cohort = FALSE,
  noaa = FALSE,
  full = FALSE,
  timeline = TRUE,
  timeline.y = 0,
  scales = "free",
  nudge.y = 20,
  ncol = 3,
  nx = 5,
  ny = 5
)
}
\arguments{
\item{...}{One or more objects of class \code{narwproj}, as returned by \link{predict.narwsim}.}

\item{diff}{Logical. If \code{TRUE}, returns plots of the differences between pairs of projections. Defaults to \code{FALSE}.}

\item{interval}{Logical. If \code{TRUE}, percentile confidence intervals are shown on the plots. Defaults to \code{TRUE}.}

\item{shade}{Logical. If \code{TRUE}, confidence intervals are displayed as colored ribbons. Otherwise, confidence bounds are shown as dotted lines. Defaults to \code{TRUE}.}

\item{cohort}{Logical. If \code{TRUE}, separate plots are returned for each cohort. If \code{FALSE}, a single plot for the whole population is shown. Defaults to \code{FALSE}.}

\item{noaa}{Logical. If \code{TRUE}, the population trajectory predicted as part of NOAA's population viability analysis (Runge et al., 2023) is also plotted. Defaults to \code{FALSE}.}

\item{full}{Logical. If \code{TRUE}, the trend plot includes historical population size estimates (dating back to 2000).}

\item{timeline}{Logical. If \code{TRUE}, the plot includes timeline(s) of offshore wind development.}

\item{timeline.y}{Numeric. Position of the timeline on the y-axis.}

\item{scales}{Character. Defines whether axis scales should be constant or vary across plots. Can be one of "fixed" or "free" (the default).}

\item{nudge.y}{Numeric. Vertical adjustment to nudge timeline labels by.}

\item{ncol}{Integer. Number of columns for the plot layout when \code{cohort = TRUE}. Defaults to \code{3}.}

\item{nx}{Integer. Desired number of x-axis intervals. Non-integer values are rounded down. Defaults to \code{5}. See \link[base]{pretty} for details.}

\item{ny}{Integer. Desired number of y-axis intervals. Non-integer values are rounded down. Defaults to \code{5}. See \link[base]{pretty} for details.}
}
\description{
Takes one or more \code{narwproj} object(s) and creates plots of the estimated population trajectories.
}
\examples{
\dontrun{
library(narwind)
m <- narw(1000)
m <- augment(m)
p <- predict(m)
plot(p)
}
}
\references{
Runge MC, Linden DW, Hostetler JA, Borggaard DL, Garrison LP, Knowlton AR, Lesage V, Williams R, Pace III RM (2023). A management-focused population viability analysis for North Atlantic right whales. US Dept Commer Northeast Fish Sci Cent Tech Memo 307, 93 p. Available at \url{https://www.fisheries.noaa.gov/s3/2023-10/TM307-508-1-.pdf}.
}
\author{
Phil J. Bouchet
}
