% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.narwsim.R
\name{predict.narwsim}
\alias{predict.narwsim}
\title{Forecast NARW abundance}
\usage{
\method{predict}{narwsim}(obj, yrs = 35, n = 1000)
}
\arguments{
\item{obj}{Object returned by narw()}

\item{yrs}{Number of years for the projection}

\item{n}{Number of replicate projections}
}
\description{
Forecast NARW abundance
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}

----------------------------
}
\author{
Phil J. Bouchet
}
