% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_backup}
\alias{predict_backup}
\title{------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
---------------------------
---------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------}
\usage{
predict_backup(
  obj,
  yrs = 35,
  n = 100,
  popr = 1,
  do.plot = FALSE,
  seed = 125897,
  ...
)
}
\description{
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
---------------------------
---------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
}
