% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_leslie}
\alias{predict_leslie}
\title{predictnarwsim <- function(obj,
                            yrs = 35,
                            n = 100) {
  
  # if(is.null(obj$gam)) stop("Insufficient data available. Cannot proceed with population projections.")
  # if(!identical(cohortID, 1:6)) stop("Missing cohorts in input <obj>. Cannot proceed with population projections.")
  # if(length(obj$gam$fit$surv$xlevels[[1]]) < 6 | length(obj$gam$fit$bc$xlevels[[1]]) < 6) stop("Missing factor levels in input <obj>. Cannot proceed with population projections.")
  
  # plogis("link" predictions + error)
  
  # Adapted from original code by Scott Creel
  # https://www.montana.edu/screel/teaching/bioe-440r-521/course-outline/stoch_projection_new.R
  # 
  # Prediction intervals
  # https://stat.ethz.ch/pipermail/r-help/2011-April/275632.html
  
  # Population estimate as per 2022 NARW report card is 340 (+/- 7).
  
  # test <- mgcv::predict.gam(mod$gam[[1]]$surv, newdata = data.frame(start_bc = seq(0,0.6, 0.01)), type = "link", se.fit = TRUE)
  # plot(seq(0,0.6, 0.01), plogis(test$fit), type = "l")
  # lines(seq(0,0.6, 0.01), plogis(Reduce("+", test)), lty = 2)
  # lines(seq(0,0.6, 0.01), plogis(Reduce("-", test)), lty = 2)
  # test2 <- mgcv::predict.gam(mod$gam[[1]]$surv, newdata = data.frame(start_bc = 0.2), type = "response")
  # abline(v = 0.2)
  # abline(h = tesct2)
  
  cohortID <- obj$param$cohortID
  cohorts <- obj$param$cohorts |> dplyr::slice(1) |> 
    dplyr::mutate(name = gsub(", female", "", name), abb = gsub(",f", "", abb)) |> 
    dplyr::bind_rows(obj$param$cohorts) |> 
    dplyr::mutate(name = gsub("male, female", "female", name), abb = gsub("m,f", "f", abb))
  cohorts <- cohorts[c(1,3,5,2,4,6,7,8)]
  
  # Attributes to monitor during projection
  mat.attribs <- c("alive", "cohort", "female", "age", "length", "length_a", "length_b", "length_c",
                   "tot_mass", "lean_mass", "bc", "mass_a", "mass_b", "p_surv", "min_bc", "calving_int", "t2calf", "birth")
  
  # Current year
  current.yr <- lubridate::year(lubridate::now())
  
  # Extract terminal functions
  mod <- obj$gam$fit
  mod[["gest"]] <- gam_gest
  
  #'------------------------------------------------------
  # INITIALIZATION
  #'------------------------------------------------------
  
  # Define initial population vector
  N0 <- c(2, 5, 212, 2, 69, 1, 7, 61)
  names(N0) <- cohorts[, name]
  
  # Create matrices and initialize them
  # rows: years <yrs>
  # columns: <attributes>
  # layers: individuals <n>
  # 4th dimension: replicate projection -> later converted to list
  narw.indiv <- array(data = NA, c(yrs + 1, length(mat.attribs), sum(N0), n), 
                      dimnames = list(paste0("yr ", 0:yrs), 
                                      mat.attribs,
                                      paste0("whale ", 1:sum(N0)),
                                      paste0("prj ", 1:n)))
  
  cohort.vec <- do.call(c, sapply(X = 1:nrow(cohorts), FUN = function(x) rep(cohorts$id[x], each = N0[x])))
  
  # Alive and population cohort
  narw.indiv[1,"alive",,] <- 1
  narw.indiv[1,"cohort",,] <- rep(cohort.vec, n)
  
  # Sex
  #  -- Calves (male)
  narw.indiv[,"female", 1:N0[1], ] <- 0  
  #  -- Calves (female)
  fem <- which(cohort.vec == 0)
  narw.indiv[,"female", fem[fem > N0[1]], ] <- 1   
  #  -- Juveniles and adults
  narw.indiv[,"female", which(cohort.vec %in% c(2,4,5,6)), ] <- 1
  narw.indiv[,"female", which(cohort.vec %in% c(1,3)), ] <- 0
  
  # Age
  ages <- start_age_vec(rep(cohort.vec, n))
  narw.indiv[1,"age", , ] <- ages
  
  # Total body length
  l.params <- agL_vec(ages)
  lengths <- age2length_vec(ages, l.params)
  narw.indiv[1,"length", , ] <- lengths
  narw.indiv[,"length_a",,] <- rep(l.params[,1], each = yrs + 1)
  narw.indiv[,"length_b",,] <- rep(l.params[,2], each = yrs + 1)
  narw.indiv[,"length_c",,] <- rep(l.params[,3], each = yrs + 1)
  
  # Total mass
  m.params <- mL(n*sum(N0))
  mass <- length2mass_vec(lengths, m.params, FALSE)
  narw.indiv[,"mass_a",,] <- rep(m.params[,1], each = yrs + 1)
  narw.indiv[,"mass_b",,] <- rep(m.params[,2], each = yrs + 1)
  narw.indiv[1,"tot_mass", , ] <- mass
  
  # Body conditon
  narw.indiv[1,"bc",,] <- start_bcondition_vec(rep(cohort.vec, n))
  
  # lean mass
  narw.indiv[1,"lean_mass", , ] <- mass - (narw.indiv[1,"bc",,] * mass)
  
  # Calving interval - mean of 5.3 years, up to apparent gaps of 13 years (Kraus et al. 2001)
  # NARWC Report Card 2022: Average inter-birth interval of 7.7 yrs, median of 8, min/max (2/13)
  # This corresponds to a Normal (7.7, 1.45)
  # Mean age at first calving = 9.53 +/- 2.32 (Kraus et al. 2001)
  # 
  # Stewart et al. 2022 -- 
  # The degree to which the energetic reserves of females are depleted during lactation may govern
  # the length of the resting period between successful pregnancies (Miller et al. 2011, Marón et al. 2015).
  narw.indiv[1,"calving_int",,] <- (rep(cohort.vec, n) == 6) * rnorm(sum(N0)*n, mean = 7.7, sd = 1.45)
  narw.indiv[1,"t2calf",,] <- round(narw.indiv[1,"calving_int",,],0)
  
  narw.indiv[1,"min_bc",,] <- predict_m(model = mod,
                                        values = as.vector(narw.indiv[1,"tot_mass",,]), 
                                        prediction = "gest") * as.vector(narw.indiv[1, "cohort",, ] == 6)
  narw.indiv[1,"birth",,] <- 0
  narw.indiv[1,"p_surv",,] <- 1
  
  #' ---------------------------
  # IMPORTANT 
  #' ---------------------------
  # Turn array into a list - otherwise cannot add new calves within projections
  narw.indiv <- purrr::array_branch(narw.indiv, 4) 
  
  # Number of individuals in each cohort
  narw.pop <- array(data = NA, c(n, yrs + 1, length(N0)), 
                    dimnames = list(paste0("prj ", 1:n),
                                    paste0("yr ", 0:yrs), 
                                    cohorts$name))
  narw.pop[,1,] <- rep(N0, each = n)
  
  # Total population size
  tot.pop <- matrix(0, n, yrs + 1, dimnames = list(paste0("prj",1:n), paste0("yr ", 0:yrs)))
  tot.pop[,1] <- sum(N0)
  
  #'------------------------------------------------------
  # RUN PROJECTIONS
  #'------------------------------------------------------
  
  # This uses nested loops. 
  # The prj loop (outermost loop) replicates the projection <n> times.
  # The i loop is next, and steps across all years of projection from an initial population vector.
  
  # Set up progress bar
  pb <- progress::progress_bar$new(
    format = "[:bar] :percent eta: :eta",
    total = n, clear = FALSE, width = 80
  )
  
  for(prj in 1:n){
    
    pb$tick() # Update progress bar
    
    #'------------------------------------------------------
    # Loop over years
    #'------------------------------------------------------
    
    for(i in 2:(yrs+1)){
      
      current.dat <- as.matrix(narw.indiv[[prj]][i-1, , ])
      alive <- narw.indiv[[prj]][i-1, "alive", ]
      
      #' ----------------------------
      # SURVIVAL
      #' ----------------------------
      # Predict survival probability based on body condition
      ps <- alive * predict_m(model = mod, cohort = current.dat["cohort",], values = current.dat["bc",], prediction = "surv")
      narw.indiv[[prj]][i, "p_surv", ] <- ps
      
      # Determine whether the animal survived
      # Maximum longevity of 69 years
      alive <- rbinom(n = dim(narw.indiv[[prj]])[3], size = 1, prob = ps) * (narw.indiv[[prj]][i-1, "age", ] <=69)
      narw.indiv[[prj]][i, "alive", ] <- alive
      
      # Sex remains the same
      narw.indiv[[prj]][i, "female", ] <- narw.indiv[[prj]][i-1, "female", ]
      
      #' ----------------------------
      # GROWTH
      #' ----------------------------
      
      # Increment age
      narw.indiv[[prj]][i, "age", ] <- alive * narw.indiv[[prj]][i-1, "age", ] + 1
      
      # Increment length
      narw.indiv[[prj]][i,"length", ] <- alive * age2length_vec(narw.indiv[[prj]][i, "age", ],
                                                                t(narw.indiv[[prj]][i, c("length_a", "length_b", "length_c"),]))
      
      # Increment lean mass
      narw.indiv[[prj]][i,"lean_mass", ] <- alive * length2mass_vec(narw.indiv[[prj]][i, "length", ],
                                                                    t(narw.indiv[[prj]][i, c("mass_a", "mass_b"),]), lean = TRUE)
      
      # Predict new body condition from current body condition
      narw.indiv[[prj]][i, "bc", ] <- alive * predict_m(model = mod, cohort = current.dat["cohort",],
                                                        values = current.dat["bc",], prediction = "bc")
      
      # Increment total mass
      narw.indiv[[prj]][i,"tot_mass", ] <- alive * narw.indiv[[prj]][i,"lean_mass", ] / (1 - narw.indiv[[prj]][i,"bc", ])
      
      #' ----------------------------
      # REPRODUCTION
      #' ----------------------------
      
      narw.indiv[[prj]][i, "calving_int", ] <- alive * narw.indiv[[prj]][i-1, "calving_int", ]
      narw.indiv[[prj]][i, "t2calf", ] <- ifelse(narw.indiv[[prj]][i-1, "t2calf", ] - 1 < 0, 0, narw.indiv[[prj]][i-1, "t2calf", ] - 1)
      
      # Minimum body condition needed to successfully bring fetus to term without starving
      # No evidence of reproductive senescence in right whales - Hamilton et al. (1998)
      narw.indiv[[prj]][i, "min_bc", ] <- alive * predict_m(model = mod, values = narw.indiv[[prj]][i, "tot_mass", ], 
                                                            prediction = "gest") * (narw.indiv[[prj]][i-1, "cohort", ] == 6)
      
      # Birth of new calf
      narw.indiv[[prj]][i, "birth", ] <- alive * (narw.indiv[[prj]][i, "bc", ] >= narw.indiv[[prj]][i, "min_bc", ]) * 
        (narw.indiv[[prj]][i-1, "cohort", ] == 6) * (narw.indiv[[prj]][i-1, "birth", ] == 0)
      
      # Maturity - transitions between cohorts
      narw.indiv[[prj]][i, "cohort", ] <- alive * increment_cohort(cohort = narw.indiv[[prj]][i-1, "cohort", ],
                                                                   age = narw.indiv[[prj]][i, "age", ],
                                                                   female = narw.indiv[[prj]][i, "female", ],
                                                                   t2calf = narw.indiv[[prj]][i, "t2calf", ])
      
      # New births
      # narw.indiv[[prj]] <- new_calf(narw.indiv[[prj]][,,], year = i, attrib = mat.attribs)
      
      # Number of animals in each cohort
      # Calves (male)
      narw.pop[prj, i, 1] <- sum((narw.indiv[[prj]][i, "cohort", ] == 0) *
                                   (narw.indiv[[prj]][i, "female", ] == 0) *
                                   (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Juveniles and adults (male)
      narw.pop[prj, i, 2] <- sum((narw.indiv[[prj]][i, "cohort", ] == 1) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 3] <- sum((narw.indiv[[prj]][i, "cohort", ] == 3) * (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Calves (female)
      narw.pop[prj, i, 4] <- sum((narw.indiv[[prj]][i, "cohort", ] == 0) *
                                   (narw.indiv[[prj]][i, "female", ] == 1) *
                                   (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Juvenile and reproductive adults (female)
      narw.pop[prj, i, 5] <- sum((narw.indiv[[prj]][i, "cohort", ] == 2) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 6] <- sum((narw.indiv[[prj]][i, "cohort", ] == 4) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 7] <- sum((narw.indiv[[prj]][i, "cohort", ] == 5) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 8] <- sum((narw.indiv[[prj]][i, "cohort", ] == 6) * (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Total population size
      tot.pop[prj, i] <- sum(narw.indiv[[prj]][i, "alive", ], na.rm = TRUE)
      
    } # End years
  } # End projections
  
  # Compile outputs
  narw.df <- purrr::map(.x = cohorts$name, .f = ~{
    narw.pop[,,.x] |> 
      tibble::as_tibble() |> 
      tibble::rownames_to_column(var = "prj") |> 
      tidyr::pivot_longer(!prj, names_to = "year", values_to = "N") |> 
      dplyr::mutate(year = current.yr + as.numeric(gsub("yr ", "", year))) |> 
      dplyr::mutate(cohort = stringr::str_to_sentence(.x))
  }) |> do.call(what = rbind) |> data.table::data.table()
  
  tot.df <- tibble::as_tibble(tot.pop) |> 
    tibble::rownames_to_column(var = "prj") |> 
    tidyr::pivot_longer(!prj, names_to = "year", values_to = "N") |> 
    dplyr::mutate(year = current.yr + as.numeric(gsub("yr ", "", year))) |> 
    dplyr::mutate(cohort = "North Atlantic right whales") |> data.table::data.table()
  
  narw.conf <- narw.df[, list(mean = mean(N), lwr = quantile(N, 0.025), uppr = quantile(N, 0.975)), list(year, cohort)] |> 
    dplyr::mutate(cohort = factor(cohort, levels = c("Calves (male)",
                                                     "Calves (female)",
                                                     "Juveniles (male)", 
                                                     "Juveniles (female)", 
                                                     "Adults (male)",
                                                     "Adults (female, pregnant)",
                                                     "Adults (female, lactating)",
                                                     "Adults (female, resting)")))
  
  tot.conf <- tot.df[, list(mean = mean(N), lwr = quantile(N, 0.025), uppr = quantile(N, 0.975)), list(year, cohort)]|> 
    dplyr::mutate(cohort = factor(cohort, levels = c("Calves (male)",
                                                     "Calves (female)",
                                                     "Juveniles (male)", 
                                                     "Juveniles (female)", 
                                                     "Adults (male)",
                                                     "Adults (female, pregnant)",
                                                     "Adults (female, lactating)",
                                                     "Adults (female, resting)",
                                                     "North Atlantic right whales")))
  
  p1 <- plot_projection(narw.df, narw.conf)
  p2 <- plot_projection(tot.df, tot.conf)
  
  print(p1)
  print(p2)
  
  # Find 95% confidence intervals on final population size
  cat("Final population size:\n")
  final.pop <- unname(tot.df[year == current.yr + yrs,  quantile(N, probs = c(0.5, 0.025, 0.975))])
  cat("N = ", final.pop[1], " (95% CI: ", final.pop[2], "–", final.pop[3], ")\n", sep = "")
  
  return(list(ind = narw.indiv, df = rbind(narw.df, tot.df), ci = rbind(narw.conf, tot.conf)))
  
}
----------------------------}
\usage{
predict_leslie(obj, n = 100, yrs = 35)
}
\description{
predictnarwsim <- function(obj,
                            yrs = 35,
                            n = 100) {
  
  # if(is.null(obj$gam)) stop("Insufficient data available. Cannot proceed with population projections.")
  # if(!identical(cohortID, 1:6)) stop("Missing cohorts in input <obj>. Cannot proceed with population projections.")
  # if(length(obj$gam$fit$surv$xlevels[[1]]) < 6 | length(obj$gam$fit$bc$xlevels[[1]]) < 6) stop("Missing factor levels in input <obj>. Cannot proceed with population projections.")
  
  # plogis("link" predictions + error)
  
  # Adapted from original code by Scott Creel
  # https://www.montana.edu/screel/teaching/bioe-440r-521/course-outline/stoch_projection_new.R
  # 
  # Prediction intervals
  # https://stat.ethz.ch/pipermail/r-help/2011-April/275632.html
  
  # Population estimate as per 2022 NARW report card is 340 (+/- 7).
  
  # test <- mgcv::predict.gam(mod$gam[[1]]$surv, newdata = data.frame(start_bc = seq(0,0.6, 0.01)), type = "link", se.fit = TRUE)
  # plot(seq(0,0.6, 0.01), plogis(test$fit), type = "l")
  # lines(seq(0,0.6, 0.01), plogis(Reduce("+", test)), lty = 2)
  # lines(seq(0,0.6, 0.01), plogis(Reduce("-", test)), lty = 2)
  # test2 <- mgcv::predict.gam(mod$gam[[1]]$surv, newdata = data.frame(start_bc = 0.2), type = "response")
  # abline(v = 0.2)
  # abline(h = tesct2)
  
  cohortID <- obj$param$cohortID
  cohorts <- obj$param$cohorts |> dplyr::slice(1) |> 
    dplyr::mutate(name = gsub(", female", "", name), abb = gsub(",f", "", abb)) |> 
    dplyr::bind_rows(obj$param$cohorts) |> 
    dplyr::mutate(name = gsub("male, female", "female", name), abb = gsub("m,f", "f", abb))
  cohorts <- cohorts[c(1,3,5,2,4,6,7,8)]
  
  # Attributes to monitor during projection
  mat.attribs <- c("alive", "cohort", "female", "age", "length", "length_a", "length_b", "length_c",
                   "tot_mass", "lean_mass", "bc", "mass_a", "mass_b", "p_surv", "min_bc", "calving_int", "t2calf", "birth")
  
  # Current year
  current.yr <- lubridate::year(lubridate::now())
  
  # Extract terminal functions
  mod <- obj$gam$fit
  mod[["gest"]] <- gam_gest
  
  #'------------------------------------------------------
  # INITIALIZATION
  #'------------------------------------------------------
  
  # Define initial population vector
  N0 <- c(2, 5, 212, 2, 69, 1, 7, 61)
  names(N0) <- cohorts[, name]
  
  # Create matrices and initialize them
  # rows: years <yrs>
  # columns: <attributes>
  # layers: individuals <n>
  # 4th dimension: replicate projection -> later converted to list
  narw.indiv <- array(data = NA, c(yrs + 1, length(mat.attribs), sum(N0), n), 
                      dimnames = list(paste0("yr ", 0:yrs), 
                                      mat.attribs,
                                      paste0("whale ", 1:sum(N0)),
                                      paste0("prj ", 1:n)))
  
  cohort.vec <- do.call(c, sapply(X = 1:nrow(cohorts), FUN = function(x) rep(cohorts$id[x], each = N0[x])))
  
  # Alive and population cohort
  narw.indiv[1,"alive",,] <- 1
  narw.indiv[1,"cohort",,] <- rep(cohort.vec, n)
  
  # Sex
  #  -- Calves (male)
  narw.indiv[,"female", 1:N0[1], ] <- 0  
  #  -- Calves (female)
  fem <- which(cohort.vec == 0)
  narw.indiv[,"female", fem[fem > N0[1]], ] <- 1   
  #  -- Juveniles and adults
  narw.indiv[,"female", which(cohort.vec %in% c(2,4,5,6)), ] <- 1
  narw.indiv[,"female", which(cohort.vec %in% c(1,3)), ] <- 0
  
  # Age
  ages <- start_age_vec(rep(cohort.vec, n))
  narw.indiv[1,"age", , ] <- ages
  
  # Total body length
  l.params <- agL_vec(ages)
  lengths <- age2length_vec(ages, l.params)
  narw.indiv[1,"length", , ] <- lengths
  narw.indiv[,"length_a",,] <- rep(l.params[,1], each = yrs + 1)
  narw.indiv[,"length_b",,] <- rep(l.params[,2], each = yrs + 1)
  narw.indiv[,"length_c",,] <- rep(l.params[,3], each = yrs + 1)
  
  # Total mass
  m.params <- mL(n*sum(N0))
  mass <- length2mass_vec(lengths, m.params, FALSE)
  narw.indiv[,"mass_a",,] <- rep(m.params[,1], each = yrs + 1)
  narw.indiv[,"mass_b",,] <- rep(m.params[,2], each = yrs + 1)
  narw.indiv[1,"tot_mass", , ] <- mass
  
  # Body conditon
  narw.indiv[1,"bc",,] <- start_bcondition_vec(rep(cohort.vec, n))
  
  # lean mass
  narw.indiv[1,"lean_mass", , ] <- mass - (narw.indiv[1,"bc",,] * mass)
  
  # Calving interval - mean of 5.3 years, up to apparent gaps of 13 years (Kraus et al. 2001)
  # NARWC Report Card 2022: Average inter-birth interval of 7.7 yrs, median of 8, min/max (2/13)
  # This corresponds to a Normal (7.7, 1.45)
  # Mean age at first calving = 9.53 +/- 2.32 (Kraus et al. 2001)
  # 
  # Stewart et al. 2022 -- 
  # The degree to which the energetic reserves of females are depleted during lactation may govern
  # the length of the resting period between successful pregnancies (Miller et al. 2011, Marón et al. 2015).
  narw.indiv[1,"calving_int",,] <- (rep(cohort.vec, n) == 6) * rnorm(sum(N0)*n, mean = 7.7, sd = 1.45)
  narw.indiv[1,"t2calf",,] <- round(narw.indiv[1,"calving_int",,],0)
  
  narw.indiv[1,"min_bc",,] <- predict_m(model = mod,
                                        values = as.vector(narw.indiv[1,"tot_mass",,]), 
                                        prediction = "gest") * as.vector(narw.indiv[1, "cohort",, ] == 6)
  narw.indiv[1,"birth",,] <- 0
  narw.indiv[1,"p_surv",,] <- 1
  
  #' ---------------------------
  # IMPORTANT 
  #' ---------------------------
  # Turn array into a list - otherwise cannot add new calves within projections
  narw.indiv <- purrr::array_branch(narw.indiv, 4) 
  
  # Number of individuals in each cohort
  narw.pop <- array(data = NA, c(n, yrs + 1, length(N0)), 
                    dimnames = list(paste0("prj ", 1:n),
                                    paste0("yr ", 0:yrs), 
                                    cohorts$name))
  narw.pop[,1,] <- rep(N0, each = n)
  
  # Total population size
  tot.pop <- matrix(0, n, yrs + 1, dimnames = list(paste0("prj",1:n), paste0("yr ", 0:yrs)))
  tot.pop[,1] <- sum(N0)
  
  #'------------------------------------------------------
  # RUN PROJECTIONS
  #'------------------------------------------------------
  
  # This uses nested loops. 
  # The prj loop (outermost loop) replicates the projection <n> times.
  # The i loop is next, and steps across all years of projection from an initial population vector.
  
  # Set up progress bar
  pb <- progress::progress_bar$new(
    format = "[:bar] :percent eta: :eta",
    total = n, clear = FALSE, width = 80
  )
  
  for(prj in 1:n){
    
    pb$tick() # Update progress bar
    
    #'------------------------------------------------------
    # Loop over years
    #'------------------------------------------------------
    
    for(i in 2:(yrs+1)){
      
      current.dat <- as.matrix(narw.indiv[[prj]][i-1, , ])
      alive <- narw.indiv[[prj]][i-1, "alive", ]
      
      #' ----------------------------
      # SURVIVAL
      #' ----------------------------
      # Predict survival probability based on body condition
      ps <- alive * predict_m(model = mod, cohort = current.dat["cohort",], values = current.dat["bc",], prediction = "surv")
      narw.indiv[[prj]][i, "p_surv", ] <- ps
      
      # Determine whether the animal survived
      # Maximum longevity of 69 years
      alive <- rbinom(n = dim(narw.indiv[[prj]])[3], size = 1, prob = ps) * (narw.indiv[[prj]][i-1, "age", ] <=69)
      narw.indiv[[prj]][i, "alive", ] <- alive
      
      # Sex remains the same
      narw.indiv[[prj]][i, "female", ] <- narw.indiv[[prj]][i-1, "female", ]
      
      #' ----------------------------
      # GROWTH
      #' ----------------------------
      
      # Increment age
      narw.indiv[[prj]][i, "age", ] <- alive * narw.indiv[[prj]][i-1, "age", ] + 1
      
      # Increment length
      narw.indiv[[prj]][i,"length", ] <- alive * age2length_vec(narw.indiv[[prj]][i, "age", ],
                                                                t(narw.indiv[[prj]][i, c("length_a", "length_b", "length_c"),]))
      
      # Increment lean mass
      narw.indiv[[prj]][i,"lean_mass", ] <- alive * length2mass_vec(narw.indiv[[prj]][i, "length", ],
                                                                    t(narw.indiv[[prj]][i, c("mass_a", "mass_b"),]), lean = TRUE)
      
      # Predict new body condition from current body condition
      narw.indiv[[prj]][i, "bc", ] <- alive * predict_m(model = mod, cohort = current.dat["cohort",],
                                                        values = current.dat["bc",], prediction = "bc")
      
      # Increment total mass
      narw.indiv[[prj]][i,"tot_mass", ] <- alive * narw.indiv[[prj]][i,"lean_mass", ] / (1 - narw.indiv[[prj]][i,"bc", ])
      
      #' ----------------------------
      # REPRODUCTION
      #' ----------------------------
      
      narw.indiv[[prj]][i, "calving_int", ] <- alive * narw.indiv[[prj]][i-1, "calving_int", ]
      narw.indiv[[prj]][i, "t2calf", ] <- ifelse(narw.indiv[[prj]][i-1, "t2calf", ] - 1 < 0, 0, narw.indiv[[prj]][i-1, "t2calf", ] - 1)
      
      # Minimum body condition needed to successfully bring fetus to term without starving
      # No evidence of reproductive senescence in right whales - Hamilton et al. (1998)
      narw.indiv[[prj]][i, "min_bc", ] <- alive * predict_m(model = mod, values = narw.indiv[[prj]][i, "tot_mass", ], 
                                                            prediction = "gest") * (narw.indiv[[prj]][i-1, "cohort", ] == 6)
      
      # Birth of new calf
      narw.indiv[[prj]][i, "birth", ] <- alive * (narw.indiv[[prj]][i, "bc", ] >= narw.indiv[[prj]][i, "min_bc", ]) * 
        (narw.indiv[[prj]][i-1, "cohort", ] == 6) * (narw.indiv[[prj]][i-1, "birth", ] == 0)
      
      # Maturity - transitions between cohorts
      narw.indiv[[prj]][i, "cohort", ] <- alive * increment_cohort(cohort = narw.indiv[[prj]][i-1, "cohort", ],
                                                                   age = narw.indiv[[prj]][i, "age", ],
                                                                   female = narw.indiv[[prj]][i, "female", ],
                                                                   t2calf = narw.indiv[[prj]][i, "t2calf", ])
      
      # New births
      # narw.indiv[[prj]] <- new_calf(narw.indiv[[prj]][,,], year = i, attrib = mat.attribs)
      
      # Number of animals in each cohort
      # Calves (male)
      narw.pop[prj, i, 1] <- sum((narw.indiv[[prj]][i, "cohort", ] == 0) *
                                   (narw.indiv[[prj]][i, "female", ] == 0) *
                                   (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Juveniles and adults (male)
      narw.pop[prj, i, 2] <- sum((narw.indiv[[prj]][i, "cohort", ] == 1) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 3] <- sum((narw.indiv[[prj]][i, "cohort", ] == 3) * (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Calves (female)
      narw.pop[prj, i, 4] <- sum((narw.indiv[[prj]][i, "cohort", ] == 0) *
                                   (narw.indiv[[prj]][i, "female", ] == 1) *
                                   (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Juvenile and reproductive adults (female)
      narw.pop[prj, i, 5] <- sum((narw.indiv[[prj]][i, "cohort", ] == 2) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 6] <- sum((narw.indiv[[prj]][i, "cohort", ] == 4) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 7] <- sum((narw.indiv[[prj]][i, "cohort", ] == 5) * (narw.indiv[[prj]][i, "alive", ] == 1))
      narw.pop[prj, i, 8] <- sum((narw.indiv[[prj]][i, "cohort", ] == 6) * (narw.indiv[[prj]][i, "alive", ] == 1))
      
      # Total population size
      tot.pop[prj, i] <- sum(narw.indiv[[prj]][i, "alive", ], na.rm = TRUE)
      
    } # End years
  } # End projections
  
  # Compile outputs
  narw.df <- purrr::map(.x = cohorts$name, .f = ~{
    narw.pop[,,.x] |> 
      tibble::as_tibble() |> 
      tibble::rownames_to_column(var = "prj") |> 
      tidyr::pivot_longer(!prj, names_to = "year", values_to = "N") |> 
      dplyr::mutate(year = current.yr + as.numeric(gsub("yr ", "", year))) |> 
      dplyr::mutate(cohort = stringr::str_to_sentence(.x))
  }) |> do.call(what = rbind) |> data.table::data.table()
  
  tot.df <- tibble::as_tibble(tot.pop) |> 
    tibble::rownames_to_column(var = "prj") |> 
    tidyr::pivot_longer(!prj, names_to = "year", values_to = "N") |> 
    dplyr::mutate(year = current.yr + as.numeric(gsub("yr ", "", year))) |> 
    dplyr::mutate(cohort = "North Atlantic right whales") |> data.table::data.table()
  
  narw.conf <- narw.df[, list(mean = mean(N), lwr = quantile(N, 0.025), uppr = quantile(N, 0.975)), list(year, cohort)] |> 
    dplyr::mutate(cohort = factor(cohort, levels = c("Calves (male)",
                                                     "Calves (female)",
                                                     "Juveniles (male)", 
                                                     "Juveniles (female)", 
                                                     "Adults (male)",
                                                     "Adults (female, pregnant)",
                                                     "Adults (female, lactating)",
                                                     "Adults (female, resting)")))
  
  tot.conf <- tot.df[, list(mean = mean(N), lwr = quantile(N, 0.025), uppr = quantile(N, 0.975)), list(year, cohort)]|> 
    dplyr::mutate(cohort = factor(cohort, levels = c("Calves (male)",
                                                     "Calves (female)",
                                                     "Juveniles (male)", 
                                                     "Juveniles (female)", 
                                                     "Adults (male)",
                                                     "Adults (female, pregnant)",
                                                     "Adults (female, lactating)",
                                                     "Adults (female, resting)",
                                                     "North Atlantic right whales")))
  
  p1 <- plot_projection(narw.df, narw.conf)
  p2 <- plot_projection(tot.df, tot.conf)
  
  print(p1)
  print(p2)
  
  # Find 95% confidence intervals on final population size
  cat("Final population size:\n")
  final.pop <- unname(tot.df[year == current.yr + yrs,  quantile(N, probs = c(0.5, 0.025, 0.975))])
  cat("N = ", final.pop[1], " (95% CI: ", final.pop[2], "–", final.pop[3], ")\n", sep = "")
  
  return(list(ind = narw.indiv, df = rbind(narw.df, tot.df), ci = rbind(narw.conf, tot.conf)))
  
}
----------------------------
}
