% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.narwsim.R
\name{print.narwsim}
\alias{print.narwsim}
\title{Overview of model results}
\usage{
\method{print}{narwsim}(obj, ...)
}
\arguments{
\item{obj}{Object returned by run_model}

\item{rows}{Number of rows to print}
}
\description{
Overview of model results
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}

}
\author{
Phil J. Bouchet
}
