% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtnorm}
\alias{rtnorm}
\title{Random deviate from a truncated Normal distribution}
\usage{
rtnorm(location, scale, L, U)
}
\arguments{
\item{location}{Location parameter}

\item{scale}{Scale parameter}

\item{L}{Lower bound}

\item{U}{Upper bound}
}
\description{
Random deviate from a truncated Normal distribution
}
