% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveParameters.R
\name{saveParameters}
\alias{saveParameters}
\title{Update scenario object when app stops and return to R environment}
\usage{
saveParameters(obj)
}
\arguments{
\item{obj}{An object of class \code{narwscenario}, as returned by \code{\link{scenario}}.}
}
\value{
An updated \code{narwscenario} object.
}
\description{
Update scenario object when app stops and return to R environment
}
\author{
Rob Schick
}
