% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{Offshore wind scenario}
\usage{
scenario()
}
\value{
An object of class \code{narwscenario}.
}
\description{
Launches an interactive Shiny application to define custom scenarios for use in the bioenergetic model.
}
\author{
Rob Schick, Phil J. Bouchet, Enrico Pirotta
}
