% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_testing.R
\name{scenario_testing}
\alias{scenario_testing}
\title{Offshore wind scenario}
\usage{
scenario_testing()
}
\value{
An object of class \code{narwscenario}.
}
\description{
Launches an interactive Shiny application to define custom scenarios for use in the bioenergetic model.
}
\author{
Rob Schick, Phil J. Bouchet, Enrico Pirotta
}
