% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{start_bcondition}
\alias{start_bcondition}
\title{Initialize body condition}
\usage{
start_bcondition(cohort)
}
\arguments{
\item{age}{Age in years}

\item{shape1}{First shape parameter of the beta distribution}

\item{shape2}{Second shape parameter of the beta distribution}
}
\description{
Performs a random draw from a beta distribution to initialize 
the body condition of simulated animals, taken as the ratio of fat mass to total mass.
}
