% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.narwproj.R
\name{summary.narwproj}
\alias{summary.narwproj}
\title{Population projection summary}
\usage{
\method{summary}{narwproj}(obj)
}
\arguments{
\item{obj}{An object of class \code{narwproj}.}
}
\description{
Prints a text-based summary of outputs from the stochastic population model to the R console.
}
\examples{
\dontrun{
library(narwind)
m <- narw(1000)
m <- augment(m)
prj <- predict(m)
summary(prj)
}
}
\author{
Phil J. Bouchet
}
