% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.scenario.R
\name{update.narwscenario}
\alias{update.narwscenario}
\title{Update scenario object}
\usage{
update.narwscenario(obj, vessel.tbl)
}
\arguments{
\item{obj}{An object of class \code{narwscenario}, as returned by \code{\link{scenario}}.}

\item{vessel.tbl}{A \code{data.frame} object defining vessel traffic parameters from custom user input to the Shiny app.}
}
\value{
An updated \code{narwscenario} object.
}
\description{
Update scenario object
}
\author{
Phil J. Bouchet
}
