% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.narwsim.R
\name{write.narwsim}
\alias{write.narwsim}
\title{Summary}
\usage{
\method{write}{narwsim}(obj, ...)
}
\description{
Summary information
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}
}
\author{
Phil J. Bouchet
}
