% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narw.R
\name{narw}
\alias{narw}
\title{Run the bioenergetic model}
\usage{
narw(nsim = 1000, scenario = NULL, n.cores = NULL, progress = TRUE, ...)
}
\arguments{
\item{nsim}{Number of simulated animals}

\item{n.prop}{Number of proposals used in the importance sampler (Michelot, 2019)}
}
\description{
Simulate right whale movements and behavior across a calendar year.
}
\examples{
\dontrun{
library(narwind)

narw(10)
}
}
\seealso{
\code{\link{initialize}}
}
\author{
Phil J. Bouchet
}
