% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narwind-package.R
\docType{package}
\name{narwind}
\alias{narwind}
\title{Population consequences of offshore wind}
\description{
The \code{narwind} package provides methods for fitting and selecting among behavioural dose-response models using Bayesian reversible jump Markov Chain Monte Carlo (rjMCMC).
}
\author{
Phil J. Bouchet
}
