% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.narwproj.R
\name{plot.narwproj}
\alias{plot.narwproj}
\title{Plot}
\usage{
\method{plot}{narwproj}(obj)
}
\arguments{
\item{obj}{Object returned by run_model}
}
\description{
Make plots
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}
}
\author{
Phil J. Bouchet
}
