% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.narwsim}
\alias{plot.narwsim}
\title{Title}
\usage{
\method{plot}{narwsim}(
  obj,
  id = NULL,
  n = 1,
  animate = FALSE,
  raster.id = NULL,
  hide.lgd = FALSE
)
}
\arguments{
\item{id}{}

\item{n}{}

\item{animate}{}

\item{raster.id}{}

\item{hide.lgd}{}

\item{animals}{}
}
\description{
Title
}
