% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.narwsim.R
\name{plot.narwsim}
\alias{plot.narwsim}
\title{Model plots}
\usage{
\method{plot}{narwsim}(obj, what = "map", animate = FALSE, web = FALSE, ...)
}
\arguments{
\item{obj}{Object returned by run_model}

\item{what}{}

\item{animate}{Logical. If TRUE, creates a .gif animation of the tracks.}

\item{web}{If TRUE, returns a plotly graph}
}
\description{
Make plots
}
\examples{
\dontrun{
library(narwind)
m <- narw(10000)
plot(m)
}
-------------------------------------------------------
-------------------------------------------------------
-------------------------------------------------------
-------------------------------------------------------  
-------------------------------------------------------
-------------------------------------------------------  
}
\author{
Phil J. Bouchet
}
