% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.narwsim.R
\name{plot.narwsim}
\alias{plot.narwsim}
\title{Plot}
\usage{
\method{plot}{narwsim}(
  obj,
  inits = FALSE,
  whaleID = NULL,
  animate = FALSE,
  web = FALSE,
  nmax = 100
)
}
\arguments{
\item{obj}{Object returned by run_model}

\item{animate}{Logical. If TRUE, creates a .gif animation of the tracks.}

\item{web}{If TRUE, returns a plotly graph}

\item{n.ind}{Number of animals to plot.}

\item{id}{ID of animals to plot}
}
\description{
Make plots
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}
}
\author{
Phil J. Bouchet
}
