% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.narwsim.R
\name{predict.narwsim}
\alias{predict.narwsim}
\title{Forecast NARW abundance}
\usage{
\method{predict}{narwsim}(obj)
}
\arguments{
\item{obj}{Object returned by run_model}
}
\description{
Forecast NARW abundance
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}

}
\author{
Phil J. Bouchet
}
