% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.narwsim.R
\name{predict.narwsim}
\alias{predict.narwsim}
\title{Forecast NARW abundance}
\usage{
\method{predict}{narwsim}(obj, n = 100, yrs = 35, ...)
}
\arguments{
\item{obj}{Object returned by narw()}

\item{n}{Number of replicate projections}

\item{yrs}{Number of years for the projection}
}
\description{
Forecast NARW abundance
}
\author{
Phil J. Bouchet

------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
------------------------------------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
----------------------------
}
