% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.narwsim.R
\name{print.narwsim}
\alias{print.narwsim}
\title{Overview of model results}
\usage{
\method{print}{narwsim}(obj, ...)
}
\arguments{
\item{obj}{An object of class \code{narwsim}, as returned by \code{\link{narw}}.}
}
\description{
Prints a subset of model data to the R console.
}
\examples{
\dontrun{
library(narwind)
m <- narw(1000)
print(m)
}

}
\author{
Phil J. Bouchet
}
