% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.narwsim.R
\name{summary.narwsim}
\alias{summary.narwsim}
\title{Summary}
\usage{
\method{summary}{narwsim}(obj)
}
\description{
Summary information
}
\examples{
\dontrun{
library(narwind)

animals <- run_model(10)
plot(animals)
}
}
\author{
Phil J. Bouchet
}
